Refactor guest and reservation status handling; streamline API calls, enhance DataTable and ModalDelete components with objName prop for improved user feedback, and update routing paths for consistency.
