+++ react_client/src/layouts/components/Sidebar.tsx
@@ -249,7 +249,7 @@ export default function Sidebar() {
             >
               <List sx={{ gap: 0.5 }}>
                 <ListItem sx={{ mt: 0.5 }}>
-                  <NavListItemButton to="/reservation-statuses">Reservation Statuses</NavListItemButton>
+                  <NavListItemButton to="/reservation-status">Reservation Status</NavListItemButton>
                 </ListItem>
                 <ListItem>
                   <NavListItemButton to="/room-types">Room Types</NavListItemButton>
